requirements.txt
pyvcf
biopython==1.81
pyfaidx==0.7.2.1
pandas==2.0.1

numpy~=1.24.3
matplotlib~=3.7.1
scikit-learn~=1.2.2
ete3~=3.1.3